+++ resources/views/admin/auth/passwords/email.blade.php
@@ -10,8 +10,18 @@
             <h1 class="text-center text-black">Je wachtwoord vergeten?</h1>
 
             <x-chief::window>
+                {{-- The status session value holds the passwords.sent message when an reset email has been succesfully requested. --}}
                 @if(session('status'))
-                    <div>{{ session('status') }}</div>
+                    <div>
+                        <p>{{ session('status') }}</p>
+                        <div class="space-x-4 mt-4">
+                            <a
+                                    href="{{ route('chief.back.login') }}"
+                                    title="Terug naar login"
+                                    class="btn btn-primary-outline"
+                            > Terug naar login </a>
+                        </div>
+                    </div>
                 @else
                     <form id="valid" role="form" method="POST" action="{{ route('chief.back.password.email') }}">
                         {{ csrf_field() }}
@@ -20,7 +30,7 @@
                             {{-- TODO: mail confirmation message also shows as error --}}
                             <x-chief::field.form error="email">
                                 <x-slot name="description">
-                                    <p>Geef je e-mailadres in om je wachtwoord te opnieuw in te stellen.</p>
+                                    <p>Geef je e-mailadres in om je wachtwoord opnieuw in te stellen.</p>
                                 </x-slot>
 
                                 <input id="identity" name="email" type="email" placeholder="E-mail" value="{{ old('email') }}">